+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add Ledger Accounts resource
 - Properly handle some project error messages
+- `access_token_expires_at` is now  set as UTC
 
 ## 1.2.1